Rework mount infrasctructure
Previously, only RootFS and RootFSObj knew anything about mounts.
This was complicated and bug-prone, and made LOOKUPP nearly impossible
to implement.  Changed to simpler system of generic mount knowledge,
with a mount hiding the directory below.  Mount traversal is done
within the object lookup and lookupp functions, and should be
transparent to the server layer.