GSS: Implement a reasonable Name proxy class
Implement a proxy class for the RFC 2743 INTERNAL NAME object that
properly handles allocation and freeing of resources, and gives
python user easy access to printable version.  There is still work to
be done integrating various gss name utility functions, but this is
a much more solid base to work off of.  Note that at the moment, the
proxy class is not returned by other gss functions.  This will be corrected
in future patches.

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>