RPC: Refactor ConnectionHandler to use single alarm socket
ConnectionHandler was using two sockets for internal alarms, one
for write pending and another for new socket pending.  It is possible
to combine these into a single socket, if use the previously passed
"garbage byte" as a command switch.  This is done primarily to simplify
upcoming addition of stop() method.

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>