Fix server conversion of connection info to pipe
Fix a bug in server introduced by commit 4b908d8da.  It was trying to
convert stored connection information into a usable communication pipe.
However, the information sent by the RPC layer was changed by a
previous commit, resulting in errors.  So we change the info passed yet
again, to just the pipe directly.

Note this means the pipe is being used as a key in a dictionary, and
that it is being stored outside of the RPC layer.  I'm not sure this
is a good idea, and may have to eventually revert to the original
fd and timestamp data that was being sent.

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>