Support automated server reboots for reboot tests
Modify _waitForReboot to call an external script to reboot the server,
specified by the user with a --rebootscript parameter.

To allow the user to write one script that may boot multiple test
servers depending on the test run, we allow the user to pass a string to
--rebootargs which will be passed as a parameter to the given reboot
script.

We also provide a sample reboot script that uses libvirt's virsh to
reboot a virtual guest.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Fred Isaman <iisaman@netapp.com>