SRVR: Rewrite DSDevice.load()
There was no reason to use a csv reader given that we weren't using
a csv format.  So just use standard open, and in the process make
the config file a lot friendlier, allowing whitespace and comments.

We also extend the format of the config file from server:port to
server[:[port][/path]].  Path isn't used yet, but that will be dealt
with in the next patch.

Signed-off-by: Fred Isaman <iisaman@netapp.com>