CLNT: Fix program number check in CSESS20
Test CSESS20 was failing and giving an error about missing attributes.
The test checks to make sure that the server sends callbacks with the
same program number as requested in the create_session. The check the
test was doing expected to be called with the entire callback message,
but was actually only being passed the program version from the message.
So to fix the test, it seems that the caller should either pass the
whole message or the check should only expect to receive the program
number. Since the caller sometimes calls other functions with the same
parameters and I don't know what those expect, as well as the fact that
the check doesn't use any part of the message other than the program
number, I wrote a quick patch that changes the check to expect only the
program number. The test now passes; patch attached.

This was due to failure to update the test after the changes made by patch
"a5569e3e RPC: Add _check_program and _version_range server hooks"

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Fred Isaman <iisaman@netapp.com>