SRVR: Prevent GETATTR from adjusting file position
GETATR(size) was determining size by doing a seek.  Unfortunately, this
can race with WRITE (which relies on a seek to set the write offset)
causing data corruption.

This patch fixes the immediate issue for file-layout, but the whole
concept of fattr4_size needs to be re-examined.  In particular, we
probably need a way to determine size without reference to seek.

We also should look at whether we can reduce GETATTR calls between MDS
and DS if internal calls to _size could differentiate based on how
strict the consistency of size needs to be between MDS and DS.

Signed-off-by: Fred Isaman <iisaman@netapp.com>