TESTS: fix DELEG9
The new server doesn't have a recall callback setup, so will never send
the delegreturn.  (Doing it by hand at the end as attempted here doesn't
work well, as the open that precedes that final delegreturn needs the
delegreturn before it can succeed.)

Also, we need to set c.cb_server for _verify_cb_occurred to get the
correct count.

After this patch, the test passes against the Linux server.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>