Generalize reboot helper script
We allow the user to pass a script on the testserver commandline that
reboots the server.

There will be more cases like this where we want to do something special
on the server that can't be done through the protocol and may depend on
local configuration details.

So instead of adding a new helper script for each one, use a single
"serverhelper" script that can be passed different commands.  Fall back
on requesting manual intervention in the absence of the script.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>