Don't spin trying to read closed socket forever
recv on a closed socket returns '' and '' is not None.

Without this patch, I can easily make a pNFS setup with two DSs spin -
using much of the CPU while doing nothing.  A quick way to reproduce: do a
hard reset (like a crash) on a client that has mounted the export -- the
connection "goes away", but the pynfs server never gives up on it.

Signed-off-by: Weston Andros Adamson <dros@netapp.com>