4.0 server tests: fix OPEN29
Note the close was using the wrong owner.  Also, I was getting spurious
DELAY's thanks to a delegation given out on the first open.  I think
that was probably due to reusing an owner from a previous test; use a
different owner name making it less likely the server gives out a
delegation.  (Note the real fix would be probably be to return any
delegation we're given here.)

Signed-off-by: J. Bruce Fields <bfields@redhat.com>