4.0 server tests: all combinations of 5 open/downgrade sequences
OK, this is an overly exhaustive test of a narrow sequence of
operations.  We should come up with something more general (e.g.
including some sequences expected to generate an error, including
different stateowners, etc.), and figure out a way to pick out just a
few of the interesting ones.

Anyway.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>