4.1 server tests: allow SYMLINK error on rename
Yes, 5661 does say to return NOTDIR in this case.  But I can't figure
out why.  Elsewhere it's generally preferred to return the more specific
SYMLINK error.  This is a buggy-client case (since the source and target
dirs are referred to by filehandle, the client should generally know
whether it's dealing with a directory or not), so I don't see the point
of being really picky here.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>