4.1 server tests: don't insist on 0-seqid trick in setup
The 4.1 spec requires servers to handle stateid's with seqid field 0
(they have the special meaning of "use the latest stateid in this
stateid stream").

We've seen buggy or incomplete servers that don't handle this yet.

We *should* add a test for that.  But insisting on it (for no real
reason) during setup just prevents all the other tests from running
unnecessarily.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>