4.1 server tests: test close with stateid.seqid = 0
Before f435d212d9553add548720975e720d7cd5ae8eba, pynfs used to do a
close with stateid.seqid = 0 routinely in setup.  That seems the wrong
place to do it, but it is supposed to work (and we have seen servers
fail to do this), so we may as well have a test for it.

Also, I mistakenly commented out that line in the previous fix when I
meant to remove it completely.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>