4.0 server tests: REBT5,6 should expect BAD not CONFLICT
The rfc 3530 language is rather skimpy.  Dave Noveck went into some more
detail in a May 2009 message to the ietf list:

	http://old.nabble.com/NFS4ERR_REC-LAIM_BAD-td23765973.html

My understanding is that RECLAIM_BAD was originally intended for the
case where a the reclaim does not match the server's stable storage
information.  Whereas RECLAIM_CONFLICT would happen if e.g. a client's
reclaim conflicts with a lock already reclaimed by another client on
this boot cycle.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>