4.1 server tests: EID6e should allow NFS4ERR_PERM
Yes, we're trying to do an update with the wrong verifier.  But we've
also got the principal wrong.  Strictly speaking this falls only under
case 8 of the spec, but I can't see why responding as in case 9 wouldn't
also be acceptable.  (And arguably perhaps there's some benefit to
performing the actual permission check earlier.)

Signed-off-by: J. Bruce Fields <bfields@redhat.com>