Add a more explicit copyright notice
There is incomplete copyright information scattered through the files,
and the setup.py files all make it clear that the license is GPL, so I
don't think there's any doubt that the whole can be distributed under
the GPL (though unidentified portions, e.g., those copyright the UM or
CITI, may also be available under more liberal terms).

But I think we should make this clearer if only to avoid any
misunderstandings with future contributors.

So add a COPYRIGHT file with a copy of GPLv2 and a brief summary of the
copyright status.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>