4.1 server tests: better REP_TOO_BIG test
We're depending on the server returning TOO_BIG on a 500-byte read over
a channel with maximum reply size 500.

It would be legal for the server to do that, but as far as I can tell it
would also be legal for it to return OK and a short read.

(And in fact depending on the server implementation that's quite likely
in this case since the file is actually only 160 bytes long!)

Use a getattrs of the mandatory attributes instead, as that reply has a
well-defined minimum size.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>