4.1 server tests: don't require large-ACL support
Some servers or filesystems aren't going to support ACLs with hundreds
of entries, and that's OK, we shouldn't be erroring out because they
don't.

Better would be to test a range of sizes and allow either OK or (past a
certain point) an appropriate error (FBIG?).

Ideal might be to actually allow pynfs to support server limits such as
this without reporting them as errors.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>