4.1 server tests: fix CSESS20 callback handling
On receiving CREATE_SESSION, Ganesha is sending a CB_NULL, waiting for a
response, and then dropping the CREATE_SESSION without reply if that
response is an error.

That behavior is a Ganesha bug.  But it also looks like the error reply
to the CB_NULL was a pynfs bug.

It might still be interesting to test the server's handling of failed
callbacks.  (After all, this seems to have been useful to uncover a
Ganesha bug.)  But that would probably be better done in a separate test
designed for that purpose.  Ideally such a test would find a more
realistic way to fail.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>