4.0 server tests: skip test expecting STALE error
Somebody else on the server could in theory be holding a reference on
this file even if it existed only briefly.  Even absent that I don't
think the linux dcache, for example, will necessarily guarantee that
this file is gone the instant the close returns.  (I'm seeing NFS_OK's
on the PUTFH here and haven't investigated why.)

Maybe something else (a delay? a reboot?) would make this reliable, or
we could just allow NFS_OK.  For now, just skip the test by default.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>