4.0 server tests: replace WRT5 by better tests
This test is sending a write that (at least in the knfsd case) is larger
than the server's maximum write size, and expecting it to succeed.
That's weird.

Remove WRT5 and replace it by two tests.  Both first query the maximum
write size.  One does a large (but not too large) write and checks that
the data was written correctly.  The other does a too-large write and
expects nothing.  (A wide variety of server behavior would be in spec;
we do this just in case a buggy server might crash.)

Signed-off-by: J. Bruce Fields <bfields@redhat.com>