Add LOCK24 test case to test open uprgade/downgrade scenario
The Ganesha server has an interesting issue with open uprgade
and downgrade while a lock stateid is active. This was actually seen
with NFS v4.1, however, this test case easily triggers the bug.

Basically, we open the file read only, take a read lock, then
unlock it. If this were 4.1, we could have a FREE_STATEID, however
the condition arises when a new open on the file occurs for write
only and a write lock is taken, all of that racing with the
FREE_STATEID and winning. Since the issue occurs with the
FREE_STATEID losing the race, we don't need FREE_STATEID at
all, so just do this in 4.0. The RFC doesn't specifically
state whether lock with new lock owner is allowed in 4.0 when a
stateid already exists (it's definitely needed in 4.1 to resolve
the FREE_STATEID/LOCK race, again, if FREE_STATEID lost the race,
the old lock stateid still exists, while if FREE_STATEID won the
race it's gone and a new one is necessary. Thus the client is
safest to assume a new one is needed and that the server will
accept that.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>