python3: need binary array instead of t.code string
Lots of tests use t.code as a stateowner of filename or other string.
But t.code is a unicode string, and probably has to be.  And we want
byte arrays, not strings, at the protocol level.

So instead, for these uses, define a t.word() function which returns a
byte array based on t.code.  You can also optionally pass an integer if
you need more than one byte array.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>