Document high-port requirement for server testing
Some NFS servers by default require the client to connect from a
low-numbered port.  But pynfs is meant to be runnable as a non-root
user, and as such can't necessarily get low-numbered ports.

It might be useful to at least try to get a low-numbered port, or to
give a better error message.  For now, at least warn about this in the
README.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>