DELEG22: grant delegs on write opens too
Even when the client has write opens, the server can still grant
delegations.

(Didn't add the "all" flag since this behavior is optional.  Then again,
so is most delegation-granting behavior.  Some day we should think about
how to report optional behavior so that we can still for example
identify regressions in a given implementation even when the behavior
we're regressing to isn't strictly incorrect.)

Signed-off-by: J. Bruce Fields <bfields@redhat.com>