rpc: use getaddrinfo instead of guessing ipv4/v6
We're guessing that anything with a ":" in it will use IPv6, and
anything else will use IPv4.

That fails if we try to test against somehost.example.com where
somehost.example.com only has an IPv6 address.

Instead, look up the given address with getaddrinfo first.

Partly lifted from cpython's create_connection implementation.  Not
completely sure this is right, but it seems to be an improvement....

Signed-off-by: J. Bruce Fields <bfields@redhat.com>