build: convert from deprecated distutils to setuptools
According to PEP 632 [1], distutils is obsolete and removed after Python
v3.12. Setuptools is the replacement.

There is a migration guide at [2] that suggests the following
replacements:

    {distutils.core => setuptools}.setup
    {distutils => setuptools}.command
    distutils.dep_util => setuptools.modified

Prior to setuptools v69.0.0 [3], `newer_group` was exposed through
the now-deprecated `setuptools.dep_util` instead of
`setuptools.modified`.

Rather than updating distutils.core.Extension, I remove it as it does
not appear to be used.

Link: https://peps.python.org/pep-0632/ [1]
Link: https://setuptools.pypa.io/en/latest/deprecated/distutils-legacy.html [2]
Link: https://setuptools.pypa.io/en/latest/history.html#v69-0-0 [3]
Signed-off-by: Thomas Bertschinger <tahbertschinger@gmail.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Tested-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Signed-off-by: Calum Mackay <calum.mackay@oracle.com>